Move keyboard_gui to utils folder
